requirements.txt
langchain
langchain_core
langchain_community
langgraph
langchain_cohere
langchain_groq
chroma
langchain_chroma
tiktoken
python-dotenv
gradio
bs4